+++ src/plans/debug.jl
@@ -101,14 +101,32 @@ function status_summary(dst::DebugSolverState)
         for (k, v) in dst.debugDictionary
             s = "$s\n    :$k = $(status_summary(v))"
         end
-        return "$(dst.state)\n\n## Debug$s"
+        return "$(status_summary(dst.state))\n\n## Debug$s"
     else # for length 1 the group is equivalent to the summary of the single state
         return status_summary(dst.state)
     end
 end
-function show(io::IO, dst::DebugSolverState)
-    return print(io, status_summary(dst))
+# 2-argument show, used by Array show, print(obj) and repr(obj), keep it short
+function Base.show(io::IO, obj::DebugSolverState)
+    return print_object(io, obj, multiline = false)
 end
+
+# the 3-argument show used by display(obj) on the REPL
+function Base.show(io::IO, ::MIME"text/plain", obj::DebugSolverState)
+    # you can add IO options if you want
+    multiline = get(io, :multiline, true)
+    return print_object(io, obj, multiline = multiline)
+end
+
+function print_object(io::IO, dst::DebugSolverState; multiline::Bool)
+    if multiline
+        return print(io, status_summary(dst))
+    else
+        # write something short, or go back to default mode
+        Base.show_default(io, dst)
+    end
+end
+
 dispatch_state_decorator(::DebugSolverState) = Val(true)
 
 #